back-16 correct types

+++ hh-dep-monitoring-service/src/main/resources/scripts/create_relation.sql
@@ -6,8 +6,8 @@ CREATE TYPE priority AS ENUM (
 
 CREATE TABLE IF NOT EXISTS relation (
     relation_id SERIAL PRIMARY KEY,
-    repository_from_id INT REFERENCES repository NOT NULL,
-    repository_to_id INT REFERENCES repository NOT NULL,
+    repository_from_id BIGINT REFERENCES repository NOT NULL,
+    repository_to_id BIGINT REFERENCES repository NOT NULL,
     description TEXT,
     priority priority NOT NULL
 );
