+++ hh-dep-monitoring-service/src/main/resources/scripts/create_repository.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS repository (
-    repository_id SERIAL PRIMARY KEY,
+    repository_id BIGINT PRIMARY KEY,
     name VARCHAR(110) NOT NULL,
     html_url VARCHAR(255) NOT NULL,
     description VARCHAR(500),
